Move Record data to record subdirectory and add create functions
